refactor: 랜덤값 생성 클래스들의 패키지 위치 변경
