+++ src/main/java/domain/random/RandomValueGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.random;
 
 public interface RandomValueGenerator {
     int generate();
